project/requirements.txt
arrow==1.2.1
asgiref==3.4.1
beautifulsoup4==4.10.0
binaryornot==0.4.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cloudinary==1.26.0
cookiecutter==1.7.3
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.5
django-admin-volt==0.0.3
django-bootstrap4==21.1
django-braces==1.15.0
django-heroku==0.3.1
django-oauth-toolkit==1.5.0
django-on-heroku==1.1.2
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.4
gcloud==0.17.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
Jinja2==3.0.3
jinja2-time==0.2.0
jwcrypto==1.0
jws==0.1.3
MarkupSafe==2.0.1
oauth2client==3.0.0
oauthlib==3.1.1
Pillow==8.4.0
poyo==0.5.0
protobuf==3.19.1
psycopg==3.0.4
psycopg2==2.9.2
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-jwt==2.0.1
python-slugify==5.0.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.3.1
sqlparse==0.4.2
stripe==2.59.0
text-unidecode==1.3
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.3
